requirements1.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
conllu==6.0.0
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz#sha256=8588289f761488021e096e8a7ebbdad52acc0e74685e602c5d0d0b7a5fa546a3
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
murmurhash==1.0.12
namex==0.0.8
narwhals==1.34.1
nmslib-metabrainz==2.1.3
numba==0.61.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pybind11==2.13.6
pydantic==1.10.21
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pysbd==0.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
scispacy
seaborn==0.13.2
shap==0.47.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart-open==6.4.0
smmap==5.0.2
spacy
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.44.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
thinc==8.1.12
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.7.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wasabi==0.10.1
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
